FIXED - bug 194734: [Databinding] Property-based observables https://bugs.eclipse.org/bugs/show_bug.cgi?id=194734
LEEROY JENKINS!!!

+++ tests/org.eclipse.jface.tests.databinding/src/org/eclipse/core/tests/databinding/conversion/StringToNumberConverterTest.java
@@ -112,7 +112,7 @@ public class StringToNumberConverterTest extends TestCase {
 		assertEquals("Non-integer BigDecimal", input, result);
 
 		// Test 2: Long
-		input = new BigDecimal((long) (Integer.MAX_VALUE + 100));
+		input = new BigDecimal(Integer.MAX_VALUE + 100L);
 		result = (BigDecimal) converter.convert(formatBigDecimal(input));
 		assertEquals("Integral BigDecimal in long range", input, result);
